requirements.txt
-Pip kullanıcıları için bağımlılık listesi, Eğitim/test için tüm kütüphaneleri içermeli.

# Temel
diffusers>=0.30.0          # 0.35.2
transformers>=4.45.0       # 4.57.1
accelerate>=1.0.0          # 1.10.1
safetensors>=0.4.5         # 0.6.2
einops>=0.8.0              # 0.8.1
opencv-python>=4.10.0.84   # 4.10.0.84
numpy>=1.26.4              # 1.26.4
scipy>=1.13.1              # 1.13.1
tqdm>=4.66.4               # 4.67.1
PyYAML>=6.0.2              # 6.0
python-dotenv>=1.0.1       # 1.1.1

# Kontrol/adapter eko-sistemi 
huggingface_hub>=0.25.2    # 0.34.3
httpx>=0.27.0              # 0.28.1

# Video I/O
imageio>=2.35.1            # 2.19.3 (DÜŞÜK)
imageio-ffmpeg>=0.5.1      # YOK ŞUAN
moviepy>=1.0.3             # YOK ŞUAN


